chore: Remove lsql-parser and lsql-tokenizer crates
