--- crates/lsql-parser/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "lsql-parser"
-version = "0.0.0"
-
-[dependencies]
-lsql-tokenizer = { path = "../lsql-tokenizer", version = "*"}
