--- crates/lsql-tokenizer/Cargo.toml
@@ -1,5 +1,0 @@
-[package]
-name = "lsql-tokenizer"
-version = "0.0.0"
-
-[dependencies]
